Update README.md
Questions preceded by a * are optional. Although they can be skipped without any deduction, it is important to know and understand the results they contain.

Ex. 1 — Time vs. space

The goal of this exercise is to consider the best available hardware and compare the feasibility of heavy computation in terms of (i) time and (ii) memory.

    As of June 2015 the fastest supercomputer publicly known is called NUDT Tianhe-2. Its speed is 33.86 PFLOPS and its storage is 12.4 PB1.

    As of August 2015 one of the fastest CPU for desktop computer is the Intel Core i7-5775R Processor which has four core running at maximal frequency of 3.8 GHz2.

    As of August 2015 the largest hard drive is almost 16TB3.

        How long would it take to perform 264 operations on NUDT Tianhe-2? What about 280 opera-tions?

        How many desktop computers would be necessary to perform 264 operations in no more than a day. What about 280 operations in no more than a month?

        How many hard drives would be necessary to store 264 bits. What about 280 bits?